refactor: reorganise date functions indashboard_chart.py
