+++ frappe/utils/dateutils.py
@@ -7,8 +7,8 @@ import frappe.defaults
 import datetime
 from frappe.utils import get_datetime
 from frappe.utils import add_to_date, getdate
-from frappe.utils.data import get_last_day_of_week
-from frappe.desk.doctype.dashboard_chart.dashboard_chart import get_period_ending
+from frappe.utils.data import get_first_day, get_first_day_of_week, get_quarter_start, get_year_start,\
+	get_last_day, get_last_day_of_week, get_quarter_ending, get_year_ending
 from six import string_types
 
 # global values -- used for caching
@@ -103,3 +103,59 @@ def get_dates_from_timegrain(from_date, to_date, timegrain="Daily"):
 			date = get_period_ending(add_to_date(dates[-1], years=years, months=months, days=days), timegrain)
 		dates.append(date)
 	return dates
+
+def get_from_date_from_timespan(to_date, timespan):
+	days = months = years = 0
+	if timespan == "Last Week":
+		days = -7
+	if timespan == "Last Month":
+		months = -1
+	elif timespan == "Last Quarter":
+		months = -3
+	elif timespan == "Last Year":
+		years = -1
+	elif timespan == "All Time":
+		years = -50
+	return add_to_date(to_date, years=years, months=months, days=days,
+		as_datetime=True)
+
+def get_period(date, interval='Monthly'):
+	date = getdate(date)
+	months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"]
+	return {
+		'Daily': date.strftime('%d-%m-%y'),
+		'Weekly': date.strftime('%d-%m-%y'),
+		'Monthly': str(months[date.month - 1]) + ' ' + str(date.year),
+		'Quarterly': 'Quarter ' + str(((date.month-1)//3)+1) + ' ' + str(date.year),
+		'Yearly': str(date.year)
+	}[interval]
+
+def get_period_beginning(date, timegrain):
+	as_str = True
+	if timegrain == 'Daily':
+		pass
+	elif timegrain == 'Weekly':
+		date = get_first_day_of_week(date, as_str=as_str)
+	elif timegrain == 'Monthly':
+		date = get_first_day(date, as_str=as_str)
+	elif timegrain == 'Quarterly':
+		date = get_quarter_start(date, as_str=as_str)
+	elif timegrain == 'Yearly':
+		date = get_year_start(date, as_str=as_str)
+
+	return date
+
+def get_period_ending(date, timegrain):
+	date = getdate(date)
+	if timegrain == 'Daily':
+		pass
+	elif timegrain == 'Weekly':
+		date = get_last_day_of_week(date)
+	elif timegrain == 'Monthly':
+		date = get_last_day(date)
+	elif timegrain == 'Quarterly':
+		date = get_quarter_ending(date)
+	elif timegrain == 'Yearly':
+		date = get_year_ending(date)
+
+	return getdate(date)
